TravelAssistant/translation_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic>=2.3.0
typing-extensions>=4.6.1
google-cloud-translate==3.11.1
google-cloud-texttospeech==2.14.1
google-cloud-storage==2.10.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0 